+++ x-pack/platform/plugins/shared/cases/public/components/cases_redesign/all_cases/all_cases_list.tsx
@@ -20,7 +20,7 @@ import { useCasesColumns } from './hooks/use_cases_columns';
 import { CasesTableFilters } from './components/table_filters';
 import { CASES_TABLE_PER_PAGE_VALUES } from './types';
 import { CasesTable } from '../../all_cases/table';
-import { CasesList } from './components/cases_list';
+import { CasesList } from './components/list_view';
 import { VIEW_TOGGLE_TABLE_ID, type ViewToggleId } from './constants';
 import { useCasesContext } from '../../cases_context/use_cases_context';
 import { CasesMetrics } from './components/cases_metrics';
@@ -30,9 +30,9 @@ import { useBulkGetUserProfiles } from '../../../containers/user_profiles/use_bu
 import { useGetCurrentUserProfile } from '../../../containers/user_profiles/use_get_current_user_profile';
 import { getAllPermissionsExceptFrom, isReadOnlyPermissions } from '../../../utils/permissions';
 import { useIsLoadingCases } from '../../all_cases/use_is_loading_cases';
-import { useAllCasesState } from './hooks/use_all_cases_state';
+import { useAllCasesState } from '../../all_cases/use_all_cases_state';
 import { useAvailableCasesOwners } from '../../app/use_available_owners';
-import { useCasesColumnsSelection } from './hooks/use_cases_columns_selection';
+import { useCasesColumnsSelection } from '../../all_cases/use_cases_columns_selection';
 import { useListFieldsSelection } from './hooks/use_list_fields_selection';
 import { useViewMode } from './hooks/use_view_mode';
 import { DEFAULT_CASES_TABLE_STATE } from '../../../containers/constants';
