+++ x-pack/platform/plugins/shared/cases/public/components/cases_redesign/all_cases/components/list_view/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { CasesList } from './cases_list';
+export { CaseListItem } from './case_list_item';
